[BUGFIX] Remove superfluous keys for crop variants.
The Method Builder->addCropVariant() adds the whole configuration array from the CropVariants.yaml to the final crop variant, which might brake the whole crop functionality in the BE, if coverAreas and focusArea are configured, too. This bugfix removes those superfluous keys/arrays before the configuration is assigned to the crop variant.