[!!!][TASK] Remove localization helper magic
From now on an integrator must take care of an existing
string in a specific XLF file. There is no check anymore
and no fallback to the XLF file ext ext:cropvariantsbuilder
anymore.

This change is necessary due to an upstream change in the
TYPO3 core.

Cause: The TYPO3 LanguageService/-Factory can not be used
in this state of the TYPO3 bootstrapping process.

Related: #18, #27